✨ Expand the mesh to ensure vertices are placed on all terrains
The Perlin distribution rarely places vertices high up, which is undesired. Instead, we'd like to always have some points placed on the highest terrace. To enforce that, "lift" all vertices so at least one of them matches the maximum height. The strength of the expansion depends on the the factor between the maximum height and the highest vertex.