♻️ Use custom terrace heights as slice heights, not just to place terraces
I didn't like the current implementation because it would just move terraces up/down based on custom heights, but they were still equally placed apart. I'd like an implementation that allowed the user to place terraces right after each other, to give an "erosion" kind of vibe to sudden height changes.
So instead of using the custom heights just to move terraces, it's used to calculate the actual slicing heights. That gives the outcome I was looking for.