🥅 Limit even further persistence and lacunarity values
In order to keep documentation and behavior consistent, I decided that octaves should always decrease in intensity (amplitude) and increase in frequency. Therefore, persistence must be greater than 0 but smaller than 1, and lacunarity must be greater than one. But these are only relevant when the number of octaves is higher than 1.
This commit contains:
- The changes necessary to enforce these new constraints.
- Updated documentation.
- Updated automated tests.