⚡️ Optimize ListExtensions.Combine
Instead of using managed code, rely on `NativeList` APIs to perform the operations faster, while still keeping allocations low.
In order to use the `NativeList` API, the return type had to be changed to `NativeList` (there were no inter-types APIs that allowed implementation without sacrifices). Some `TerracedMeshBuilder` code had to be changed for that, but as the code comment explains, to no big performance or allocation impact.