🤡 Improve Display sample startup time
The previous implementation was warming up all terrain setups on start, which would always leave the application hanging for a few seconds.
To fix that, this commit introduces a flag that controls whether the warm-up will happen on startup for all terrains of if it will happen on-the-fly for each setup individually, whenever the previous terrain setup is being displayed.